requirements.txt
# --- Web crawling ---
requests
beautifulsoup4
feedparser
lxml

# --- Data processing ---
pandas
numpy

# --- NLP & ML ---
scikit-learn
sentence-transformers
underthesea

# --- Visualization ---
matplotlib
seaborn
wordcloud

# --- Database ---
psycopg2-binary

# --- Config ---
python-dotenv

# --- App ---
fastapi
uvicorn[standard] 
pydantic